+++ src/javascript/binary/common_functions/attach_dom/password_strength_meter.js
@@ -1,4 +1,5 @@
 var testPassword = require('../passwordmeter').testPassword;
+var CommonFunctions = require('../common_functions').CommonFunctions;
 
 var PasswordMeter = (function(){
     'use strict';
@@ -9,7 +10,7 @@ var PasswordMeter = (function(){
      * @param id               optional id for meter element
      */
     function attach($container) {
-        if (isIE()) return;
+        if (CommonFunctions.isIE()) return;
         var $meter = $('<meter></meter>', {id: 'password-meter', min: 0, max: 50, high: 20, low: 10, optimum: 50});
         $container
             .append($meter);
