--- src/javascript/binary/common_functions/common_data.js
@@ -1,7 +1,0 @@
-var CommonData = {
-    getLoginToken: function() { return Cookies.get('login'); }
-};
-
-module.exports = {
-    CommonData: CommonData,
-};
