+++ src/javascript/binary/static_pages/home.js
@@ -1,0 +1,21 @@
+var submit_email = require('../websocket_pages/user/verify_email').submit_email;
+
+var Home = (function() {
+    var init = function() {
+        check_login_hide_signup();
+        submit_email();
+    };
+    var check_login_hide_signup = function() {
+        if (page.client.is_logged_in) {
+            $('#verify-email-form').remove();
+            $('.break').attr('style', 'margin-bottom:1em');
+        }
+    };
+    return {
+        init: init,
+    };
+})();
+
+module.exports = {
+    Home: Home,
+};
