+++ src/javascript/binary/websocket_pages/mb_trade/mb_contract.js
@@ -1,5 +1,6 @@
 var objectNotEmpty = require('../../base/utility').objectNotEmpty;
 var format_currency = require('../../common_functions/currency_to_symbol').format_currency;
+var japanese_client = require('../../common_functions/country_base').japanese_client;
 var MBDefaults = require('./mb_defaults').MBDefaults;
 var MBSymbols = require('./mb_symbols').MBSymbols;
