+++ src/javascript/binary/websocket_pages/resources/asset_index/asset_indexws.ui.js
@@ -1,7 +1,9 @@
 var showLoadingImage = require('../../../base/utility').showLoadingImage;
 var Table = require('../../../common_functions/attach_dom/table').Table;
+var CommonFunctions = require('../../../common_functions/common_functions').CommonFunctions;
 var AssetIndexData = require('./asset_indexws.data').AssetIndexData;
 var AssetIndex = require('../asset_indexws').AssetIndex;
+var japanese_client = require('../../../common_functions/country_base').japanese_client;
 
 var AssetIndexUI = (function() {
     "use strict";
@@ -66,7 +68,7 @@ var AssetIndexUI = (function() {
 
         if (isFramed) {
             $container.find('ul').hide();
-            $('<div/>', {class: 'center-text'}).append(jqueryuiTabsToDropdown($container)).prependTo($container);
+            $('<div/>', {class: 'center-text'}).append(CommonFunctions.jqueryuiTabsToDropdown($container)).prependTo($container);
         }
     };
