+++ src/javascript/binary/websocket_pages/resources/market_times/market_timesws.data.js
@@ -1,3 +1,5 @@
+var japanese_client = require('../../../common_functions/country_base').japanese_client;
+
 var MarketTimesData = (function() {
     "use strict";
