+++ src/javascript/binary/websocket_pages/resources/market_times/market_timesws.ui.js
@@ -1,5 +1,7 @@
 var showLoadingImage = require('../../../base/utility').showLoadingImage;
 var Table = require('../../../common_functions/attach_dom/table').Table;
+var CommonFunctions = require('../../../common_functions/common_functions').CommonFunctions;
+var japanese_client = require('../../../common_functions/country_base').japanese_client;
 var MarketTimesData = require('./market_timesws.data').MarketTimesData;
 var MarketTimes = require('../market_timesws').MarketTimes;
 
@@ -76,7 +78,7 @@ var MarketTimesUI = (function() {
 
         if (isFramed) {
             $container.find('ul').hide();
-            $('<div/>', {class: 'center-text'}).append(jqueryuiTabsToDropdown($container)).prependTo($container);
+            $('<div/>', {class: 'center-text'}).append(CommonFunctions.jqueryuiTabsToDropdown($container)).prependTo($container);
         }
     };
