+++ src/javascript/binary/websocket_pages/socket.js
@@ -2,10 +2,13 @@ var getSocketURL = require('../../config').getSocketURL;
 var getAppId = require('../../config').getAppId;
 var Login = require('../base/login').Login;
 var objectNotEmpty = require('../base/utility').objectNotEmpty;
-var CommonData = require('../common_functions/common_data').CommonData;
+var CommonFunctions = require('../common_functions/common_functions').CommonFunctions;
+var AccountOpening = require('../common_functions/account_opening').AccountOpening;
 var SessionDurationLimit = require('../common_functions/session_duration_limit').SessionDurationLimit;
+var checkClientsCountry = require('../common_functions/country_base').checkClientsCountry;
 var Cashier = require('../websocket_pages/cashier/cashier').Cashier;
 var PaymentAgentWithdrawWS = require('../websocket_pages/cashier/payment_agent_withdrawws').PaymentAgentWithdrawWS;
+var create_language_drop_down = require('../common_functions/attach_dom/language_dropdown').create_language_drop_down;
 
 /*
  * It provides a abstraction layer over native javascript Websocket.
@@ -102,7 +105,7 @@ function BinarySocketClass() {
         }
 
         binarySocket.onopen = function () {
-            var apiToken = CommonData.getLoginToken();
+            var apiToken = CommonFunctions.getLoginToken();
             if (apiToken && !authorized && localStorage.getItem('client.tokens')) {
                 binarySocket.send(JSON.stringify({authorize: apiToken}));
             } else {
@@ -221,7 +224,7 @@ function BinarySocketClass() {
                         page.contents.topbar_message_visibility('show_residence');
                     }
                     if (/realws|maltainvestws|japanws/.test(window.location.href)) {
-                        handle_account_opening_settings(response);
+                        AccountOpening.handle_account_opening_settings(response);
                     }
                     GTM.event_handler(response.get_settings);
                     page.client.set_storage_value('tnc_status', response.get_settings.client_tnc_status || '-');
