+++ src/javascript/binary/websocket_pages/trade/charts/highchartws.js
@@ -1,5 +1,6 @@
 var MBTradePage = require('../../mb_trade/mb_tradepage').MBTradePage;
 var MBContract  = require('../../mb_trade/mb_contract').MBContract;
+var japanese_client = require('../../../common_functions/country_base').japanese_client;
 
 var Highchart = (function() {
   var chart, options, chart_forget, responseID, contract, contract_ended, contracts_for_send, history_send, entry_tick_barrier_drawn, initialized, chart_delayed, chart_subscribed, request, min_point, max_point, start_time, purchase_time, now_time, end_time, entry_tick_time, is_sold, sell_time, sell_spot_time, is_settleable, exit_tick_time, exit_time;
