+++ src/javascript/binary/websocket_pages/trade/common.js
@@ -1,5 +1,6 @@
 var objectNotEmpty = require('../../base/utility').objectNotEmpty;
 var format_money = require('../../common_functions/currency_to_symbol').format_money;
+var japanese_client = require('../../common_functions/country_base').japanese_client;
 
 /*
  * This contains common functions we need for processing the response
