+++ src/javascript/binary/websocket_pages/trade/tradepage.js
@@ -1,4 +1,5 @@
 var Guide = require('../../common_functions/guide').Guide;
+var japanese_client = require('../../common_functions/country_base').japanese_client;
 
 var TradePage = (function(){
