+++ src/javascript/binary/websocket_pages/user/account/settings/iphistory.js
@@ -1,3 +1,5 @@
+var japanese_client = require('../../../../common_functions/country_base').japanese_client;
+
 pjax_config_page_require_auth("user/security/iphistoryws", function(){
     return {
         onLoad: function() {
