+++ src/javascript/binary/websocket_pages/user/account/statement/statement.ui.js
@@ -3,6 +3,7 @@ var downloadCSV = require('../../../../base/utility').downloadCSV;
 var Button = require('../../../../common_functions/attach_dom/button').Button;
 var Table = require('../../../../common_functions/attach_dom/table').Table;
 var showTooltip = require('../../../../common_functions/get_app_details').showTooltip;
+var japanese_client = require('../../../../common_functions/country_base').japanese_client;
 
 var StatementUI = (function(){
     "use strict";
