+++ src/javascript/binary/websocket_pages/user/new_account/financial_acc_opening/financial_acc_opening.ui.js
@@ -98,7 +98,7 @@ var FinancialAccOpeningUI = (function(){
     ValidAccountOpening.checkDate(elementObj['dobdd'], elementObj['dobmm'], elementObj['dobyy'], errorObj['dobdd']);
     ValidAccountOpening.checkPostcode(elementObj['postcode'], errorObj['postcode']);
 
-    if (elementObj['residence'].value === 'gb' && /^$/.test(Trim(elementObj['postcode'].value))){
+    if (elementObj['residence'].value === 'gb' && /^$/.test((elementObj['postcode'].value).trim())){
       errorPostcode.innerHTML = Content.errorMessage('req');
       Validate.displayErrorMessage(errorPostcode);
       window.accountErrorCounter++;
@@ -112,7 +112,7 @@ var FinancialAccOpeningUI = (function(){
 
     for (key in elementObj){
       if (elementObj[key].offsetParent !== null && key !== 'address2' && key !== 'postcode' && key !== 'state') {
-        if (/^$/.test(Trim(elementObj[key].value)) && elementObj[key].type !== 'checkbox'){
+        if (/^$/.test((elementObj[key].value).trim()) && elementObj[key].type !== 'checkbox'){
           errorObj[key].innerHTML = Content.errorMessage('req');
           Validate.displayErrorMessage(errorObj[key]);
           window.accountErrorCounter++;
