+++ src/javascript/binary/websocket_pages/user/new_account/japan_acc_opening.js
@@ -1,4 +1,6 @@
 var ValidAccountOpening = require('../../../common_functions/valid_account_opening').ValidAccountOpening;
+var CommonFunctions = require('../../../common_functions/common_functions').CommonFunctions;
+var AccountOpening = require('../../../common_functions/account_opening').AccountOpening;
 
 pjax_config_page("new_account/japanws", function(){
   return {
@@ -9,8 +11,8 @@ pjax_config_page("new_account/japanws", function(){
         window.location.href = page.url.url_for('trading');
         return;
       }
-      handle_residence_state_ws();
-      detect_hedging($('#trading-purpose'), $('.hedging-assets'));
+      AccountOpening.handle_residence_state_ws();
+      CommonFunctions.detect_hedging($('#trading-purpose'), $('.hedging-assets'));
       $('#japan-form').submit(function(evt) {
         evt.preventDefault();
         if (JapanAccOpeningUI.checkValidity()){
