+++ src/javascript/binary/websocket_pages/user/new_account/japan_acc_opening/japan_acc_opening.ui.js
@@ -107,13 +107,13 @@ var JapanAccOpeningUI = function () {
       }
     }
 
-    if (/[`~!@#$%^&*)(_=+\[}{\]\\\/";:\?><,|\d]+/.test(Trim(elementObj['fname'].value))) {
+    if (/[`~!@#$%^&*)(_=+\[}{\]\\\/";:\?><,|\d]+/.test((elementObj['fname'].value).trim())) {
       errorObj['fname'].innerHTML = Content.errorMessage('reg', [letters, space, hyphen, period, apost]);
       Validate.displayErrorMessage(errorObj['fname']);
       window.accountErrorCounter++;
     }
 
-    if (/[`~!@#$%^&*)(_=+\[}{\]\\\/";:\?><,|\d]+/.test(Trim(elementObj['lname'].value))) {
+    if (/[`~!@#$%^&*)(_=+\[}{\]\\\/";:\?><,|\d]+/.test((elementObj['lname'].value).trim())) {
       errorObj['lname'].innerHTML = Content.errorMessage('reg', [letters, space, hyphen, period, apost]);
       Validate.displayErrorMessage(errorObj['lname']);
       window.accountErrorCounter++;
@@ -151,7 +151,7 @@ var JapanAccOpeningUI = function () {
 
     for (key in elementObj) {
       if (elementObj[key].offsetParent !== null && key !== 'address2') {
-        if (/^$/.test(Trim(elementObj[key].value)) && elementObj[key].type !== 'checkbox') {
+        if (/^$/.test((elementObj[key].value).trim()) && elementObj[key].type !== 'checkbox') {
           errorObj[key].innerHTML = Content.errorMessage('req');
           Validate.displayErrorMessage(errorObj[key]);
           window.accountErrorCounter++;
