+++ src/javascript/binary/websocket_pages/user/new_account/real_acc_opening/real_acc_opening.ui.js
@@ -64,7 +64,7 @@ var RealAccOpeningUI = (function(){
     ValidAccountOpening.checkDate(elementObj['dobdd'], elementObj['dobmm'], elementObj['dobyy'], errorObj['dobdd']);
     ValidAccountOpening.checkPostcode(elementObj['postcode'], errorObj['postcode']);
 
-    if (elementObj['residence'].value === 'gb' && /^$/.test(Trim(elementObj['postcode'].value))){
+    if (elementObj['residence'].value === 'gb' && /^$/.test((elementObj['postcode'].value).trim())){
       elementObj['postcode'].innerHTML = Content.errorMessage('req');
       Validate.displayErrorMessage(elementObj['postcode']);
       window.accountErrorCounter++;
@@ -76,7 +76,7 @@ var RealAccOpeningUI = (function(){
 
     for (key in elementObj){
       if (elementObj[key].offsetParent !== null && key !== 'address2' && key !== 'postcode' && key !== 'state') {
-        if (/^$/.test(Trim(elementObj[key].value)) && elementObj[key].type !== 'checkbox'){
+        if (/^$/.test((elementObj[key].value).trim()) && elementObj[key].type !== 'checkbox'){
           errorObj[key].innerHTML = Content.errorMessage('req');
           Validate.displayErrorMessage(errorObj[key]);
           window.accountErrorCounter++;
