+++ src/javascript/binary/websocket_pages/user/view_popup/view_popupws.js
@@ -3,6 +3,7 @@ var toJapanTimeIfNeeded = require('../../../base/utility').toJapanTimeIfNeeded;
 var objectNotEmpty = require('../../../base/utility').objectNotEmpty;
 var format_money = require('../../../common_functions/currency_to_symbol').format_money;
 var format_money_jp = require('../../../common_functions/currency_to_symbol').format_money_jp;
+var japanese_client = require('../../../common_functions/country_base').japanese_client;
 var MBPrice = require('../../mb_trade/mb_price').MBPrice;
 var MBTradePage = require('../../mb_trade/mb_tradepage').MBTradePage;
